move struct to .cc file and update README
